requirements.txt
pdfminer.six
tabula-py
pandas
numpy
django
unipath
django-environ
django-crispy-forms